backend/video_clipping/requirements.txt
fastapi
uvicorn
twelvelabs
boto3
ffmpeg-python
python-multipart
requests
boto3-stubs[s3]
supabase
python-dotenv